requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-community
langchain-text-splitters
pypdf
sentence-transformers
faiss-cpu